updates for Wednesday
Fixed the problem with multiple date time stamps in directory.
Most recent config is read and now compared against the last known config and stored only if there are changes and or if the config never existed.

If the file exists and no changes in the config are apparent it skips the storing of the config.

The compliance report only shows one compliance audit per device ongoing for every recent  device config in the store.